fixup cyclic deps
